Improve string safety and clarify player control logic
Replaced more uses of strcpy/strncpy/strcat with safer string handling functions (uqm::strncpy_safe and related utilities) across the codebase to prevent buffer overflows. Local string buffers are now zero-initialized. Refactored string copying to use spans and explicit length calculations for robustness. Updated PlayerControlFlags comments to clarify control modes, and zero-initialized the PlayerControl array. Refactored SetPlayerInput to use testFlag and added logging for control methods. Changed sound decoder initialization log from error to info.

+++ src/libs/graphics/gfxload.cpp
@@ -202,7 +202,7 @@ void* _GetCelData(uio_Stream* fp, uqm::DWORD length)
 			// zipped ani file
 			if (n)
 			{
-				uqm::strncpy_safe(aniDirName, {_cur_resfile_name, n - 1});
+				uqm::strncpy_safe(aniDirName, {_cur_resfile_name, static_cast<uint32_t>(n - 1)});
 				aniFileName = _cur_resfile_name + n;
 			}
 			else
@@ -222,7 +222,7 @@ void* _GetCelData(uio_Stream* fp, uqm::DWORD length)
 		else
 		{
 			// unpacked ani file
-			uqm::strncpy_safe(filename, {_cur_resfile_name, n});
+			uqm::strncpy_safe(filename, {_cur_resfile_name, static_cast<uint32_t>(n)});
 			aniFile = fp;
 			aniDir = contentDir;
 		}
@@ -421,7 +421,7 @@ void* _GetFontData(uio_Stream* fp, uqm::DWORD length)
 				s1 = s2;
 			}
 			n = s1 - _cur_resfile_name + 1;
-			uqm::strncpy_safe(fontDirName, {_cur_resfile_name, n - 1});
+			uqm::strncpy_safe(fontDirName, {_cur_resfile_name, static_cast<uint32_t> (n - 1)});
 			fontZipName = _cur_resfile_name + n;
 		}
